Split build graph core impl
